refactor: move get attendance records
